Restore test with gates
