+++ qutip/tests/core/test_superop_reps.py
@@ -14,8 +14,10 @@ from qutip import (
     Qobj, basis, identity, sigmax, sigmay, qeye, create, rand_super,
     rand_super_bcsz, rand_dm_ginibre, tensor, super_tensor, kraus_to_choi,
     to_super, to_choi, to_kraus, to_chi, to_stinespring, operator_to_vector,
-    vector_to_operator, sprepost, destroy, swap
+    vector_to_operator, sprepost, destroy
 )
+from qutip.core.gates import swap
+
 
 tol = 1e-8
