+++ qutip/tests/core/test_tensor.py
@@ -1,10 +1,14 @@
 import numpy as np
 
 from numpy.testing import assert_equal, assert_
-
+import pytest
+import functools
+import itertools
+import qutip
 from qutip import (
     Qobj, identity, sigmax, to_super, to_choi, rand_super_bcsz, basis,
-    tensor_contract, tensor_swap, num, QobjEvo, destroy, tensor
+    tensor_contract, tensor_swap, num, QobjEvo, destroy, tensor,
+    expand_operator
 )
 
 
@@ -137,3 +141,117 @@ def test_tensor_and():
     assert tensor(evo, sx)(t) == (evo & sx)(t)
     assert tensor(sx, evo)(t) == (sx & evo)(t)
     assert tensor(evo, evo)(t) == (evo & evo)(t)
+
+
+def _permutation_id(permutation):
+    return str(len(permutation)) + "-" + "".join(map(str, permutation))
+
+
+def _tensor_with_entanglement(all_qubits, entangled, entangled_locations):
+    """
+    Create a full tensor product when a subspace component is already in an
+    entangled state.  The locations in `all_qubits` which are the entangled
+    points in the output are ignored and can take any value.
+
+    For example,
+        _tensor_with_entanglement([|a>, |b>, |c>, |d>], (|00> + |11>), [0, 2])
+    should product a tensor product like (|0b0d> + |1b1d>), i.e. qubits 0 and 2
+    in the final output are entangled, but the others are still separable.
+
+    Parameters:
+        all_qubits: list of kets --
+            A list of separable parts to tensor together.  States that are in
+            the locations referred to by `entangled_locations` are completely
+            ignored.
+        entangled: tensor-product ket -- the full entangled subspace
+        entangled_locations: list of int --
+            The locations that the qubits in the entangled subspace should be
+            in in the final tensor-product space.
+    """
+    n_entangled = len(entangled.dims[0])
+    n_separable = len(all_qubits) - n_entangled
+    separable = all_qubits.copy()
+    # Remove in reverse order so subsequent deletion locations don't change.
+    for location in sorted(entangled_locations, reverse=True):
+        del separable[location]
+    # Can't separate out entangled states to pass to tensor in the right places
+    # immediately, so tensor in at one end and then permute into place.
+    out = qutip.tensor(*separable, entangled)
+    permutation = list(range(n_separable))
+    current_locations = range(n_separable, n_separable + n_entangled)
+    # Sort to prevert later insertions changing previous locations.
+    insertions = sorted(zip(entangled_locations, current_locations),
+                        key=lambda x: x[0])
+    for out_location, current_location in insertions:
+        permutation.insert(out_location, current_location)
+    return out.permute(permutation)
+
+
+def _apply_permutation(permutation):
+    """
+    Permute the given permutation into the order denoted by its elements, i.e.
+        out[0] = permutation[permutation[0]]
+        out[1] = permutation[permutation[1]]
+        ...
+
+    This function is its own inverse.
+    """
+    out = [None] * len(permutation)
+    for value, location in enumerate(permutation):
+        out[location] = value
+    return out
+
+
+class Test_expand_operator:
+    @pytest.mark.parametrize(
+        'permutation',
+        tuple(itertools.chain(*[
+            itertools.permutations(range(k)) for k in [2, 3, 4]
+        ])),
+        ids=_permutation_id)
+    def test_permutation_without_expansion(self, permutation):
+        base = qutip.tensor([qutip.rand_unitary(2) for _ in permutation])
+        test = expand_operator(base, permutation, [2] * len(permutation))
+        expected = base.permute(_apply_permutation(permutation))
+        np.testing.assert_allclose(test.full(), expected.full(), atol=1e-15)
+
+    @pytest.mark.parametrize('n_targets', range(1, 5))
+    def test_general_qubit_expansion(self, n_targets):
+        # Test all permutations with the given number of targets.
+        n_qubits = 5
+        operation = qutip.rand_unitary(2**n_targets, dims=[[2]*n_targets]*2)
+        for targets in itertools.permutations(range(n_qubits), n_targets):
+            expected = _tensor_with_entanglement([qutip.qeye(2)] * n_qubits,
+                                                 operation, targets)
+            test = expand_operator(operation, targets, [2] * 5)
+            np.testing.assert_allclose(test.full(), expected.full(),
+                                       atol=1e-15)
+
+    def test_cnot_explicit(self):
+        test = expand_operator(qutip.gates.cnot(), [2, 0], [2]*3).full()
+        expected = np.array([[1, 0, 0, 0, 0, 0, 0, 0],
+                             [0, 0, 0, 0, 0, 1, 0, 0],
+                             [0, 0, 1, 0, 0, 0, 0, 0],
+                             [0, 0, 0, 0, 0, 0, 0, 1],
+                             [0, 0, 0, 0, 1, 0, 0, 0],
+                             [0, 1, 0, 0, 0, 0, 0, 0],
+                             [0, 0, 0, 0, 0, 0, 1, 0],
+                             [0, 0, 0, 1, 0, 0, 0, 0]])
+        np.testing.assert_allclose(test, expected, atol=1e-15)
+
+    @pytest.mark.parametrize('dimensions', [
+        pytest.param([3, 4, 5], id="standard"),
+        pytest.param([3, 3, 4, 4, 2], id="standard"),
+        pytest.param([1, 2, 3], id="1D space"),
+    ])
+    def test_non_qubit_systems(self, dimensions):
+        n_qubits = len(dimensions)
+        for targets in itertools.permutations(range(n_qubits), 2):
+            operators = [qutip.rand_unitary(dimension) if n in targets
+                         else qutip.qeye(dimension)
+                         for n, dimension in enumerate(dimensions)]
+            expected = qutip.tensor(*operators)
+            base_test = qutip.tensor(*[operators[x] for x in targets])
+            test = expand_operator(base_test, targets=targets, dims=dimensions)
+            assert test.dims == expected.dims
+            np.testing.assert_allclose(test.full(), expected.full())
